# server/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sanchiva API",
    "description": "Sanchiva, Everything that matters, one place. REST API for expenses, loans, monetary data, and events.\n\n## Authentication\nMost endpoints require a JWT **Bearer** access token.\n\n1. Call **POST /api/auth/register** or **POST /api/auth/login** (email) or **POST /api/auth/guest** (or sign in via Google OAuth in the app) to get `access_token`.\n2. Click **Authorize** (lock icon, top right).\n3. Paste the access token only (Swagger adds the word Bearer for you).\n4. Click **Authorize** → **Close**, then try protected endpoints.\n\n**Developed by Ujjwal Gupta**",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign up with email and password",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "confirm_password": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "name": "Ujjwal",
                "email": "user@example.com",
                "password": "secret123",
                "confirm_password": "secret123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created + tokens"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already registered"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with email and password",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "email": "user@example.com",
                "password": "secret123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens issued"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh access token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refresh_token"
                ],
                "properties": {
                  "refresh_token": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "refresh_token": "your-refresh-token"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New access and refresh tokens"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Current authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout (revokes refresh token; wipes guest data)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste your JWT access token (from POST /api/auth/guest or login). Do not include the word \"Bearer\"."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}